requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
python-dotenv==0.21.0